+++ gtk2_ardour/tracker/midi_track_toolbar.cc
@@ -644,7 +644,7 @@ MidiTrackToolbar::show_existing_midi_automations ()
 	const std::set<Evoral::Parameter> params = midi_track->midi_playlist ()->contained_automation ();
 	for (std::set<Evoral::Parameter>::const_iterator p = params.begin (); p != params.end (); ++p) {
 		Grid::IndexParamBimap::right_const_iterator it = grid.col2params[track_index].right.find (*p);
-		size_t column = (it == grid.col2params[track_index].right.end ()) || (it->second == 0) ?
+		int column = (it == grid.col2params[track_index].right.end ()) || (it->second == 0) ?
 			grid.add_midi_automation_column (track_index, *p) : it->second;
 
 		// Still no column available, skip
@@ -659,10 +659,10 @@ MidiTrackToolbar::show_existing_midi_automations ()
 void
 MidiTrackToolbar::hide_midi_automations ()
 {
-	std::set<size_t> to_remove;
-	for (std::set<size_t>::iterator it = grid.visible_automation_columns.begin ();
+	std::set<int> to_remove;
+	for (std::set<int>::iterator it = grid.visible_automation_columns.begin ();
 	     it != grid.visible_automation_columns.end (); ++it) {
-		size_t column = *it;
+		int column = *it;
 		Grid::IndexParamBimap::left_const_iterator c2p_it = grid.col2params[track_index].left.find (column);
 		if (c2p_it == grid.col2params[track_index].left.end ()) {
 			continue;
@@ -675,7 +675,7 @@ MidiTrackToolbar::hide_midi_automations ()
 			to_remove.insert (column);
 		}
 	}
-	for (std::set<size_t>::iterator it = to_remove.begin ();
+	for (std::set<int>::iterator it = to_remove.begin ();
 	     it != to_remove.end (); ++it)
 		grid.visible_automation_columns.erase (*it);
 }
